// apps/web/src/api/generated/model/nodeDetails.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ImageNet Taxonomy API
 * OpenAPI spec version: 1.0.0
 */
import type { NodeAncestor } from './nodeAncestor'

export interface NodeDetails {
  id: number
  label: string
  path: string
  size: number
  childCount: number
  depth: number
  index: number
  ancestors: NodeAncestor[]
}